use enum for result status to capture partial success
Currently, the eclearing backend will return the status
"ok" if it has accepted at least one of the records we're
sending. It could be, though, that the rest has been refused.
We want to at least log those cases so we need to distinguish
this status from the complete success status.